ability to add components (#1)
updatePageFile now rewrites the return statement of the page component
based on the new page state. Unfortuantely it doesn't look like the ReturnStatement
node lets you set it's structure, so I think we have to set it manually as a string.

AddComponentButton is used to add new components to the page through the front end.

prettier is used because code formatting is hard. It slows down the write speed a bit,
though it's definitely not slow. Probably there are more lightweight tools and better
ways to do this.

Added unit tests for updatePageFile. This also fixes a limitation of updatePageFile
where it wasn't able to add new props to a component, only modify preexisting ones.