Add HexColor prop type (#3)
Add `HexColor` as a special type that users can import for props. It uses the built-in HTML color picker that returns a hexadecimal color value.

J=SLAP-2216
TEST=auto, manual

See that a color picker is added and using it sets the component props. The initial value is also properly seeded from the index file.

<img width="972" alt="Screen Shot 2022-07-13 at 12 56 08 PM" src="https://user-images.githubusercontent.com/88398086/178789171-154876f9-ecb4-4a19-955c-d111163f6f2a.png">