Add right click context menu to editor  (#7)
Users can now right click added components in the LHS editor to show a custom context menu with a delete option.

J=SLAP-2243
TEST=manual

Test that on right click, a custom context menu is displayed with a delete option.  When the delete option is pressed, the context menu goes away, and the component is deleted.  On clicking the `update` button, the prop is removed from the `index.tsx` file.