moved opening browser to plugin to prevent manual reload (#17)
Used `buildStart` hook to open the browser window for development. In the original item, we wanted to remove the "localhost refused to load" screen, but it doesn't seem possible since loading components depends on accessing the site- therefore the site must be opened in `buildStart` and not `load`. With this, we've at least removed the need to manually reload the window after loading finishes.

Another behavior change is that saving the file opens the browser too, because the plugin reruns on every `build` call.

J=SLAP-2287