add non template string streams props support
Renames the current StreamsDataPath to StreamsTemplateString,
and adds a new StreamsDataPath that does not allow for interpolating.
This is for things like numbers.

Adds the KG hitchhiker community icon for both, with a different green tint
for the template string version

Also adds an error boundary, so that passing a StreamsDataPath that has the incorrect
typing and causes a component to error out doesn't crash the whole app.

TEST=manual

test that I can cause the error boundary to pop up by passing in the entire streams document
instead of a number, and that the error boundary will go away once I start typing again