use discriminated unions + enum for PropState, refactor models (#21)
This PR refactors the models to use discriminated unions + enums for things like PropState, instead of passing around hardcoded string values everywhere. Originally, this PR was for auto-generating stream configs, however that has been pushed to a following PR. I kept parseObjectLiteralExpression because I ended up using it in a few places to simplify code.

PropState also now includes the type of the prop, to act as the discriminator. This removes the need for the very annoying moduleNameToComponentMetadata lookup and the error handling that entailed due to things being possibly undefined.

In general the models needed to be refactored/rethought about a bit and I feel more comfortable with things after this.

TEST=manual

can read from file and write to file