auto generate stream config (#22)
This PR adds the ability to auto generate the stream config when writing to a page file.
We do not support accessing the stream document using bracket notation, and only support dot notation.
I.e. we support `document.id` but not `document['id']` or `document["id"]`. We may need to add bracket notation in the future to support fields with things like periods and dashes in their names. (We may also not need to, since afaik custom fields are formatted like `c_myCustomField`, and for general developer use this doesn't seem too important)

We currently do not validate that the document fields used in studio actually exist. I plan to experiment with some kind of autocomplete/validation stuff in future PR. (I also need to address some organizational TODOs)

TEST=manual,auto

test that I can update the page file to point to various pieces of streams data and the config's fields will be automatically generated