autocomplete UI for streams local data (#27)
This commit adds autocomplete UI for streams props. I tried experimenting with libraries like monaco-editor,
react-select, and react-simple-code-editor + prismjs, but they either did not do what we wanted
(autocomplete), had too many features that would clutter up the UI (monaco, react-select), or just
plain did not work, or at least were very difficult to use (react-select, it seems like it's only really good
for a static list of options?)
 
We should properly use SASS or maybe try harder to tailwindify stuff, not sure which direction to go right now
so just left it as inline style.

TEST=manual

autocomplete works for both StreamsString and StreamsData. can hit enter to autocomplete stuff and it feels similar to coding in vscode. ctrl + z does NOT work for the autocompleted stuff, couldn't find a solution but it shouldn't change what we have here.

autocompleting a template string places the cursor selection in the right place, rather than at the end of the input
StreamsString allows for either template string or a normal expression, StreamsData only allows for a normal expression