global component mechanism (#30)
This pr implement the mechanism for Marking Components as "Global". For now, I decided to go with file naming convention.

User would define a global component with file name `Blah.global.tsx`, and export two things:
- a default functional component `Blah`
- a `globalProps` variable that will be use whenever this `Blah` global component is use on the page

Once admin made changes to the props through Studio and hit update, `globalProps` in the global component will be updated to reflect the change.

User may either define a new component and prop interface OR they may import another component and prop interface in another file to use as long as they reexport under component the is the same as the file name. Studio will recursively look into the imports to find the props interface declaration to properly populate the editor.

J=SLAP-2299
TEST=manual&auto

see that global components are parsed and display correctly in Studio
see that the new jest tests passed