parsing support for nested components, partial UI support
This PR adds the ability toe parse nested components that use
props.children, and represent them in the UI as a tree.
Saving to file does not work.

Updated uuid usages from v1 to v4, since v1 uuids looked too similar
and were annoying to test with.

TEST=manual

check in react dev tools that the parentUUIDsFromRoot is correct