represent a ComponentState tree using a flat array
After working with https://www.npmjs.com/package/@minoru/react-dnd-treeview,
which also uses a flat array to represent a tree, I realized things would be much
easier if the tree was represented using a flat array.

This very slightly increases the complexity of traversing the tree in a
left first breadth-first way, but drastically simplifies the complexity in moving
nodes around, since we won't have to keep a node's children array in sync.
This should also make adding/removing nodes easier for the same reason.
With this data structure we can just chop off child nodes without having to
update the parent at all.

TEST=manual

nested components render on the page