represent a ComponentState tree using a flat array (#40)
After working with https://www.npmjs.com/package/@minoru/react-dnd-treeview,
which also uses a flat array to represent a tree, I realized things would be much
easier if the tree was represented using a flat array.

This very slightly increases the complexity of traversing the tree in a
left first breadth-first way, but drastically simplifies the complexity in moving
nodes around, since we won't have to keep a node's children array in sync.
This should also make adding/removing nodes easier for the same reason.
With this data structure we can just chop off child nodes without having to
update the parent at all.

Originally, I planned to add a uuid to ComponentState mapping to make lookup easier.
However, this is unnecessary and likely premature optimization, and makes it more difficult to
update the state of the page. For example if you wanted to add a component to the page, we would
have to add it to both the mapping as well as the flattened state. Instead, it's perfectly fine to just
iterate through the entire array for now. I already want to start using redux in the future since the state
management is becoming more and more complicated, and we could add it then as an optimization if we really need to.

TEST=manual

nested components render on the page