update deps (#57)
Vite and @vitejs/plugin-react is required dependencies for Studio to work so they should not be in devDep

Note: there is some issue with tsconfig linking to the top level tsconfig.base.json and default import issue with react-dom/client when doing npm pack in a separate repo. But I will make a new item for that.

TEST=manual

ran `npm pack` and installed studio in another repo. See that running studio command no longer show errors for not finding vite and  @vitejs/plugin-react
<img width="763" alt="Screen Shot 2022-11-30 at 2 01 51 PM" src="https://user-images.githubusercontent.com/36055303/204885636-9d191173-c6f5-474a-a16e-48235315ee77.png">