Studio State Manager (#62)
this PR added a Zustand store based what is described in the tech design, containing three slices: FileMetadataSlice, PageSlice, and SiteSettingSlice. Updated studio-plugin package to export its types as they are also use in the store, and will be use in the studio components later on.

Added jest tests, which is written in a format that read/write state outside of React components ([doc](https://github.com/pmndrs/zustand#readingwriting-state-and-reacting-to-changes-outside-of-components)). When use in React component, the syntax to use the store will be similar to React hook ([doc](https://github.com/pmndrs/zustand#then-bind-your-components-and-thats-it))

J=SLAP-2489
TEST=auto

see that the added jest tests passed successfully