PageFile write new PageState to sourceFile (#65)
this PR adds a method `updatePageFile` to PageFile that takes in a PageState and writes it back to file. Additionally, it takes a a config option that includes `updateStreamConfig` option to update stream config -- update "config" variable with the found stream document data paths in the new page state and add '@yext/page' import for typing. A new file `StreamConfigOperator` is added to house data parsing and updating logic for Stream.

refactored most of `StudioSourceFile` code into `StudioSourceFileParser` and `StudioSourceFileWriter` for now, and used [ts-mixer ](https://www.npmjs.com/package/ts-mixer)npm library's `Mixin` method to have `StudioSourceFile` extends `StudioSourceFileParser` and `StudioSourceFileWriter`.

Note: This PR does not contain logic to add parameter `IndexPage({ document }) {...}` to page component. There are no methods for manipulating / creating an ObjectBindingPattern at the moment (github issue: https://github.com/dsherret/ts-morph/issues/775). Edit: I found a workaround for it, will update in Module write-to-file PR

This PR also contains a fix for a bug I found -- in a nested component structure (at least 2 level deep), the `parentUUID` is incorrectly set (set to a higher parent uuid instead of the direct parent uuid). This is fix with the changes in `parseJsxChild` function.

J=SLAP-2472
TEST=auto

see that existing and new jest tests passed.