Add ParsingOrchestrator to studio-plugin (#66)
This PR adds ParsingOrchestrator to the studio-plugin, which is used for directing the parsing process in Studio.

This replaces the getFileMetadata temporary helper we had, and instead passes in the ParsingOrchestrator's own getFileMetadata. There was some circular trickiness since a Module can reference other Modules. 

In order to generate a particular ModuleMetadata (which is a kind of FileMetadata) we have to pass it getFileMetadata. But getFileMetadata works by calling ModuleFile.getModuleMetadata(), which in turns needs a getFileMetadata. 
To solve this, we first generate all ComponentMetadata, which do not getFileMetadata to be generated. For generating ModuleMetadata, when getFileMetadata encounters a new ModuleMetadata, we recursively call ModuleFile.getModuleMetadata() and getFileMetadata() until we reach a ModuleMetadata that does not reference any other modules (this is our base case).

I also separated out parseComponentState from StudioSourceFile, which is the only portion of StudioSourceFile that needs getFileMetadata. Otherwise, all usages of StudioSourceFile would have to pass a getFileMetadata, even if it was not used in any way.

J=SLAP-2490
TEST=auto