Add props editor section to Studio (#71)
This PR added the `EditorPanel` in the right hand side of the Studio preview, with `ComponentEditor` to render a list of `PropEditor` for each of the props in the component/module based on its `PropShape`, `initialProps` and its current `props` value. Each `PropEditor` also have the option to switch between `Literal` or `Expression` value input.

this PR also setup react jest testing for Studio component with a util component to mock Zustand store.

TEST=manual&auto

see that added jest tests passed successfully. Smoke test through apps/test-site.