Support adding pages (#73)
This PR adds support for adding new pages through the UI and saving them in the store via a new `addPage` action. The UI is also updated to display a list of the pages and it allows you to click on the names to switch between them. This PR also makes other changes that are described in the page operations tech design:
- removes `setPages` from the `PageSliceActions`
- fixes a bug in `setActivePageName`
- adds a new `pendingChanges` section of `PageSliceStates` with `pagesToUpdate`
- updates `setActivePageState` and `setActiveComponentProps` to update the `pagesToUpdate`

J=SLAP-2498
TEST=auto, manual

See that the new and existing Jest tests pass. In the test-site, see that clicking the button for adding a page causes a modal to appear. Only valid page names are accepted and actually create a page.

https://user-images.githubusercontent.com/88398086/209705261-2a2ce7a9-b8fb-4939-9615-2fa8ffd02775.mov