Add undo/redo functionality (#75)
Use `zundo` to add undo/redo functionality to the Zustand store. This new temporal store is available through the `useTemporalStore` hook and includes snapshots of past and future states and actions for undoing, redoing, and clearing the temporal store.

J=SLAP-2491
TEST=auto, manual

See that the added Jest tests pass. See that clicking the buttons in the test-site works as expected.

https://user-images.githubusercontent.com/88398086/210027770-c0836365-2c77-4ea0-8344-10b0cca8aa22.mov