fix preview and placeholder on drag from component tree (#82)
This PR fix an issue where preview drag shown a box span across the whole screen and didn't include the actual text (element's name). The placeholder highlight also was obstructed by the grey background on hover be the selected item so z-index is use to properly display the highlight.

I noticed it was a bit difficult to place the element as a child of another element. But I will leave that for an actual item since that seems like it would need a bit of time to investigate and improve.

J=none
TEST=manual