Remove element from page (#81)
Add support for removing an element from a page. When a container element is removed, its direct children become the direct children of the removed element's parent.

Since there was no icon in the mocks for this action, I added an X next to each element in the tree to mimic how the remove page button looks.

J=SLAP-2529
TEST=auto, manual

In the test-site, see that top-level, container, and leaf elements can be removed and the component tree is updated as expected.