Implement File System Operations (#79)
This PR setup client-server communication in studio and implement the commit changes flow. A "Save" button is added to the action bar, which will invoke commitChanges action in store on click. commitChanges will send a message to server, where FileSystemManager will delete or update the page files. Once done, the server will send a response back to the client side to display a notification that the changes were saved successfully.

Note: There's no UX mock for this flow. I implemented the "Save" button and the toast notification for now.

J=SLAP-2525 (Also addressed errors from SLAP-2538 while working on this)
TEST=manual&auto

https://user-images.githubusercontent.com/36055303/210880942-7fedba59-bce6-458c-a2ce-5d9c45a48ec8.mov