Support for creating modules (#83)
Add support for creating a module from the active component (which is now highlighted in blue) and writing to file. Renders a modal to prompt the user for a name for the module, similar to adding a new page. Created a `modulesToUpdate` chunk in the Store to track the file metadata changes for write-to-file. Updated/new modules are not previewed as part of imported components because the imported file for that module would either not exist or be out of date. Instead, the preview is rendered via `ComponentTreePreview` (previously `PagePreview`).

J=SLAP-2542
TEST=auto, manual

See that a module is created, the page preview remains the same, and write-to-file creates a new module file and updates the page file as expected.