Wrap multiple top-level elements in a Fragment (#87)
If multiple top-level elements are present in a page's component tree on write-to-file, they are wrapped in a Fragment.

J=SLAP-2545
TEST=auto, manual

Try saving a page with more than one top-level component and see that write-to-file is successful. See that if a page has only a single top-level component it is not wrapped in a Fragment.