Pin Vite to Major Version v4.
Previously, `studio.ts` was using `npx vite --config`. This will install the
latest stable version of Vite. If a new Major was released, that would be
installed. However, that Major may be incompatible with our existant Vite
Config. For safety's sake, I believe Studio should be pinned to a specific
Major.

TEST=manual

Made sure Studio still ran as expected in the test site.