Ability to Edit Modules (#92)
This PR adds the ability to edit the contents of modules. It does not add the ability to elevate props
within components of a module to the outside of a module.

A new `pages.moduleUUIDBeingEdited` state was added, which represents the ComponentState.uuid of a module that is currently being edited.

In order to support editing modules (adding, removing, and updating components) through the same UI as editing a page, logic in the studio store - for things like updating the active components props - were moved from the PageSlice to the top level. Preexisting actions, like `pages.setActiveComponentProps`, were replaced by dumber more generic ones like `pages.setComponentProps`, which is called by the new `actions.updateActiveComponentProps`.  `actions.updateActiveComponentProps` contains switching logic to handle whether the page is in module editing mode or page editing mode.

This centralizes a lot of business logic (mainly the 2 different cases of editing page vs editing module), at the downside of making the assumption we'll always want this behavior. Looking at the mocks, I do think we'll always want this behavior, though.

The ability to nickname modules (aka the module renaming flow) will be a separate PR. We need to figure out a way to store the user's nicknames which is out of scope of this PR.
Detaching modules will also be in a separate PR.

Also simplified the CommitChangesButton.

J=SLAP-2544
TEST=manual,auto

retested core functionality 
modules
* can add components to modules
* can rearrange components within modules
* can remove components from modules
* can update props within modules

pages
* can add components to pages
* can rearrange components within pages
* can remove components from pages
* can update props of components on pages