Add HMR for User Files (#107)
This PR adds hot-module-reloading support for the user's src files.
This means that when the user refreshes the page, we will now show correct data without needing
to restart the server. Yay! Note that due to how the studio virtual module only seeds the zustand store and does
not drive it, this means that the zustand store value will not directly change due to an HMR update.
This makes 0 difference to the user, assuming that the state the frontend sends the backend is consistent
with the state the backend sends back to the frontend via HMR. We can add validation in the future to handle
what happens when this isn't the case (which should be rare anyways).

For things like modules/component files, though, changes will be automatically HMR-d to the client.
In more human words, you can update the internal contents of a component, hit save, and immediately see
your changes in the client.

J=SLAP-2564
TEST=manual,auto

I don't love how my unit tests temporarily mutates the contents of the fixture, but this fixtures is only designed to be
used by a single test, so I think it's okay.
I messed around with ts-morph's in memory filesystem, as well as the option of completely injecting a custom
FileSystemHost into the ts-morph project. Both methods are more tedious and heavy handed then they are worth,
requiring either juggling an in memory filesystem and writing multiple files to it as setup, or implementing ~19 methods
of a custom file system since for some reason spreading some preexisting file system host and only overriding the methods
I wanted didn't seem to work.