Fix bugs with streams expression props (#111)
Fix some bugs related to using streams expressions as props:
- The `activeEntityFile` was only set when populating the store with the initial studio data. So, the entity file (or undefined) for the first page in the list was used, and never changed. `setActivePage` (previously `setActivePageName`) has been changed to update the `activeEntityFile` when the active page changes
- `useExpressionSources` didn't remove `document` as an expression source if `activeEntityFile` changed from being defined to undefined, so a page without an entity file would be able to use expressions powered by the previous page's entity file
- A named import for `TemplateProps` would be added to the wrong import statement when the `document` is first referenced for a page's props because we were doing a `!==` check instead of `===`

Also, updated the write-to-file logic so a stream config is only added to the page if a document path is used in the component tree. Otherwise, a stream config with empty fields is added to the page on Save even when you're not trying to do anything with stream expressions.

J=none
TEST=auto, manual

See that switching between pages with and without entity files has the correct behavior for allowing streams expression props. Writing to file adds `TemplateProps` to the `@yext/pages` import statement.