Setup for NPM packages (#108)
This PR makes the following changes to set up `studio` and `studio-plugin` to be published as public NPM packages:
- Update the package versions and remove the `private` property so that the NPM packages can be made public.
- Update `lodash` imports that didn't have recognized default exports.
- Update the vite config to allow component files to be imported when Studio is used in an external repo (not within the monorepo).

By default, the vite server restricts which files in the filesystem can be accessed from the app to the root of the workspace when valid (i.e. `package.json` has a `workspaces` field) or otherwise the project root in the config. In our monorepo, this is the top-level package.json, so any file within the monorepo can be accessed. In an external repo that doesn't have a `workspaces` field in its `package.json`, it defaults to the directory in which the vite config is located since we've set the `root` to `__dirname` in the vite config. This means files within the external repo couldn't be accessed. This is fixed by extending the `fs.allow` paths to include the current working directory that executes the script (i.e. top-level of the external repo). I also explicitly added the vite config's directory in case someone wanted to locally point to studio from the external repo without npm packing it.

J=SLAP-2570
TEST=manual

`npm pack` studio-plugin and studio and see that an external repo (created using CRA) can run Studio. Publish alpha versions of studio-plugin and studio and see that installing the npm version of studio also works.