Create modules folder if it doesn't exist (#117)
When creating a module from the UI and writing-to-file, if the `modules` folder doesn't exist, it is created so Studio doesn't throw an error when trying to write the new module file. Also, added support for nested folder structure within the `modules` directory.

J=SLAP-2582
TEST=auto, manual

See that creating a module without a `modules` folder in the repo creates the `modules` folder and writes the new module file within it. Create a module with a nested path (e.g. `src/modules/a/b/Module.tsx`) through the UI and see that the intermediate folders are also created, along with the new module file. Make sure that Studio can start up with an existing nested `modules` folder structure in the repo.