Update Store when HMR Occurs (#115)
This PR adds a custom HMR listener so that the store will update it's contents when
the StudioData is updated, for example if the user edits the siteSettings values or adds/updates initialProps
in a component.

Previously, the initialStudioData would be updated, however because the store only uses that data to seed itself
and did not listen to it for updates, it would not be able to respond to HMR updates without reseeding the store via page refresh. This PR usages of initialStudioData so that HMR syncing logic can be consolidated in one place.

Fixes and improves some of the type: module handling we recently added, previously the test-site no longer worked,
the studio config was no longer being imported, and we lost support for streams props.
Ideally I would want this in a separate PR but there were too many merge conflicts from a practical standpoint.
* changes studio.config.ts -> studio.config.js so it can be imported again
* adds type: module to test-site and sample-component so they work again
* removed JsonImporter and commonjs build of studio-plugin, allowing us to also remove index-cjs.ts
* Keeps the conditional export in studio-plugin's package.json for require() so that the studio package's jest tests still run
* adds a babel plugin for import assertions so that studio-plugin's jest tests run that need it (ParsingOrchestrator and getLocalDataMapping)
* fixed dynamic import not accessing the .default property
* removes the unnecessary --experimental-vm-modules flag from bin/studio.ts

J=SLAP-2583, SLAP-2614
TEST=manual

I created items to add acceptance tests for these. I don't believe HMR can be tested via Jest
see that I can update the Banner.tsx initial props or siteSettings.ts values and they will be reflected in the
UI, WITHOUT clobbering any local changes to the page state

streams props work again, tested with document.address.city