Preserve 'slug' field in `TemplateConfig` for PagesJS repos. (#126)
Often, when using Studio with a PagesJS repo, a template's Stream Config will be manually edited to include the `slug` field. This field allows for the use of prod URLs during development. However, each time that template was saved from Studio, the `slug` field was removed. This is because Studio wholesale replaced the `fields` attribute. 

We will now retain the `slug` in `fields` (if it is already present). I think that this PR highlights two larger things in Studio: manual dependency injection is getting difficult and we don't have good separation from PagesJS. In some ways, better DI could reduce the PagesJS coupling. We need to think about both problems and come up with a plan. But, that's for other items. 

J=SLAP-2617
TEST=auto

Updated and added unit tests.