Add ContentPanel
This PR separates the ComponentEditor component, which
rendered all props for a component, into the PropertiesPanel
(for non-string props) and the ContentPanel (for string props).

This PR also updates expression support to only be allowed for
strings, removing the OptionPicker we had for switching between
literals and expressions. Now, all strings props specified for the
UI are assumed to be template literal expressions.

Refactors the naming of related child components and their tests.

J=SLAP-2608
TEST=manual

see I can specify a template literal string prop and save it
see the content and properties panel separate props by strings and non-strings