Improve Highlighting and add ComponentName Tag to it
This PR improves the HighlightingContainer to no longer add
extra divs to the preview, which can pretty easily mess up styling
especially because we were adding our own styling to the containers
like position: relative.

Instead, this PR uses getBoundingClientRect to get the bounding rect
of the current active component, and a single absolutely positioned
div to highlight that specific area. This also makes it possible to
add the component name tags present in the mocks, so I went ahead
and figured out the styling for that as well.

A ResizeObserver is used to handle screen resizing, it seems to work
quite well (way less lag than I'm used to with resize event listeners).

To do this, I had to use findDOMNode, which is a deprecated method.
However, in the documentation for findDOMNode, this use case
(listed as "adding a wrapper div element") is the only use case
listed as not having a better equivalent, and is the reason the method
is deprecated but not removed. This leads me to think it will not be
difficult to update this class once React 19 comes out (which I doubt will
be soon anyways).
https://beta.reactjs.org/reference/react-dom/findDOMNode

I removed StrictMode because otherwise it logs a console.error about
findDOMNode. Ideally we could keep StrictMode but only when we are developing
(and not when the user is using the app).

J=SLAP-2613
TEST=manual

can highlight components and resize the pag
removing a component will remove highlighting