Fix Creating Modules (#130)
Previously, there was a bug where saving new modules and HMR would not interact properly.

When a save is finished being written to file, HMR events are queued up and fired one at a time,
one for each updated file. 
(It appears that HMR updates do not begin after all saves are finished
being written, which I believe is due to javascript being single threaded and that executeSaveChanges
is a synchronous method.)
Each HMR update was responsible for updating the StudioData with its own personal update,
before sending an HMR message to the frontend.
This would result in sending HMR events with the save only partially complete, for example telling
the frontend to update to use a new module, before telling the frontend that the module exists.

In hotReloadStore, we currently accept a few different update types (module, component, etc.) and
also a "full" type. However, updating events to always be a "full" update would not resolve this issue,
because events would still be fired off one file at a time.
Instead, I updated `executeSaveChanges` (which is called when a save occurs) to call reloadFile (which
updates the StudioData for a particular file update) after every single file update during the save.
This causes the next HMR update to have the full StudioData.

There also was an (arguable) bug in FileSystemManager where we were writing to module files two times, once through pendingChanges.modulesToUpdate and the other through FileSystemWriter.syncMetadata(). This was not causing
any issues before (just doing extra work), but after adding this reloadFile logic it was causing ts-morph to try to
access "removed or forgotten nodes", so I removed pendingChanges.modulesToUpdate since it seems like we're
leaning towards migrating away from the pendingChanges type of update management.

I created an HmrManager class since while debugging I tried adding some pausing/resuming logic for HMR updates,
as well as sending updates in a few other places. The class isn't necessary anymore but I felt like it was nice to
bundle the different we use for HMR updates into a class.

J=SLAP-2623
TEST=manual

can save modules again without the page erroring out
HMR still works, I can manually update the contents of a page file and see it reflected in the client
deleting pages and saving works