Highlighting Without Container <div> (#132)
This PR improves the HighlightingContainer to no longer add
extra divs to the preview, which can pretty easily mess up styling
especially because we were adding our own styling to the containers
like position: relative.

Instead, this PR uses getBoundingClientRect to get the bounding rect
of the current active component, and a single absolutely positioned
div to highlight that specific area. This also makes it possible to
add the component name tags present in the mocks, so I went ahead
and figured out the styling for that as well.

A ResizeObserver is used to handle screen resizing, it seems to work
quite well (way less lag than I'm used to with resize event listeners).

To do this, I had to use findDOMNode, which is a deprecated method.
However, in the documentation for findDOMNode, this use case
(listed as "adding a wrapper div element") is the only use case
listed as not having a better equivalent, and is the reason the method
is deprecated but not removed. This leads me to think it will not be
difficult to update this class once React 19 comes out (which I doubt will
be soon anyways).
https://beta.reactjs.org/reference/react-dom/findDOMNode

Currently StrictMode logs a console.error about findDOMNode. 
The error doesn't affect behavior but we may want to conditionally turn on StrictMode but only
for when we as a team are developing, but I'm not sure how we'd want to do that.

J=SLAP-2613
TEST=manual

can highlight components by clicking on either the left sidebar nodes or in the preview itself
highlighting responds to browser window resizes
removing the current active component will remove highlighting from the page
see that I can move the active component and the highlighting will also move
if a component returns a fragment, the first possible DOM node will be highlighted

<img width="750" alt="image" src="https://user-images.githubusercontent.com/23005393/218884842-fbf49f14-5978-4352-aa74-b2a678bbb3ba.png">