Edge cases for module highlighting and component insertion order (#134)
This PR addresses some edge cases for active module highlighting and component insertion order:
- When a module was the active component, it was highlighted correctly. But, the highlighting remained on the preview of the module after clicking on the Edit Module button because the `activeComponentUUID` was not being reset
- If the active component accepts children, it's not guaranteed that all of its children will come after it in the `componentTree`, so inserting a new component directly after the active component is not sufficient to ensure that it will appear as the first child of the active component

J=SLAP-2636
TEST=auto, manual

See that after clicking on Edit Module, there is no active highlighting on the new preview. Also, see that dragging the first element in the tree to be under a container and then adding a new component while the container is the active component inserts the new component directly under the container and not under the element that was previously at the top of the tree.