Fixing removing components (#133)
Fix the removal of container components. Previously, removing a container left all of its children in a detached state in the component tree because their `parentUUID` no longer existed. The UI showed the container and its children being correctly deleted, even though that technically wasn't the case in the `componentTree`.

Also, fix the test-site's example of expression props used in a module's components to pass a string instead of number, since we no longer support expressions for non-string props.

J=SLAP-2634
TEST=auto, manual

Check that removing containers from a page or a module deletes the container element and all elements in it's sub-tree.