Move 'Create Module' Button to 'Properties' Panel. (#142)
This PR makes a few changes to the `CreateModuleButton` Component. It now appears on the 'Properties' Panel instead of the 'Actions' Bar. Additionally, the styling of the button itself was changed according to Max's desired UI. 

J=SLAP-2612
TEST=manual

Verified that when a Module was active, the appropriate actions appeared in 'Properties'. Verified that when a normal Component was active, the "Module Actions" section only contained the `CreateModuleButton`.