Support Streams Document in Modules (#148)
To support streams documents in Modules, all modules created via the frontend will always have a prop interface
that accepts the stream document as a required prop, and will be automatically passed the document as a prop.
The stream document does not appear in the UI as an editable prop.

New RecordMetadata and RecordProp models were created for representing the stream document, which is of type
`Record<string, any>`. In the future, they can be expanded for more general `Record` support, but currently
are expected to only support the stream document, and have their typescript types written in a way to signify that.

Support for required properties in PropShape was also added, since the document is always a required prop inside modules.

This PR also addresses a bug (could also be argued that this is a new feature) where module props that are passed expressions would not be rendered in the preview correctly.
For example given a page like
```jsx
const document = { name: 'oliver shi' }
function Page({ document }) {
  return <Module title={document.name}/>
}

function Module(props) {
  return <Banner text={props.title}/>
}
```
In the preview, you would expect the Banner to have text 'oliver shi', however instead it would have text of 'document.name'.

I tried to write keep the stream document logic in the front end as best as I could, so that the backend
could be more generic and flexible. Right now I'm imagining the back-end's job is to parse and write react code
as exactly as possible, and the front-end can add specific business logic like "always add a stream document to modules".

-----
### Note about why I chose to have modules ALWAYS passed the stream document as a prop
At first, I included some logic that would try to only add the document and pass it as a prop when necessary, since
modules can exist that don't need the document. However, I opted to not add this logic because 
1. it was relatively complex (requiring bottom-up tree traversals and mutating the component tree)
2. there's not much stopping us from adding it in the future; it's compatible with the current code and more of a refinement
3. I couldn't identify much perceivable gain from adding it now

For instance, assume we have some page with a module C that is nested inside a module B that is nested inside a module A, none of which use the stream document. 
Next, the user will use the field picker UI (to be added later) in order to use the stream document inside module C.
We will now need to figure out some way pass the stream document all the way from the page, to A, to B, and finally to C.
There are two ways we can do this, Context, or prop drilling.

Context would be a nice way to do this, but would require Studio to export a react context, and also have the user properly set up this context with their site. We would also need to add code (the useContext call) directly to the body of the module, instead of just updating the return statement, which is not something Studio is currently capable of.

With prop drilling, we have to somehow know both that A and B need to be updated to have their prop interfaces include the document, and that all usages of A, B, and C on every single page and parent module they are called in now must be passed in a stream document.

J=SLAP-2598
TEST=manual,auto

created a new module Heyo. created another module Yayo that included both Heyo and ContainerWithButtons (another module) as internal components. Saw that saving the file resulted in the expected files.